requirements.txt
requests~=2.24.0
pycryptodome~=3.14.1